+++ CHANGES.md
@@ -9,6 +9,7 @@
     -   Element.toc -> Element.get_toc()
     -   Element.text_changes -> Element.get_text_changes()
     -   Element.tracked_changes -> Element.get_tracked_changes()
+    -   Element.user_defined_list -> Element.get_user_defined_list()
 
 ### Added
 
@@ -17,6 +18,7 @@
 -   Added `Element.toc`
 -   Added `Element.text_changes`
 -   Added `Element.tracked_changes`
+-   Added `Element.user_defined_list`
 
 ## [3.7.9] - 2024-05-3